Bug 1176337 - Remove unnecessarily gcc-4.4-specific cast from nsTableRowFrame::ReflowChildren(), now that we require newer gcc. r=dholbert

+++ layout/tables/nsTableRowFrame.cpp
@@ -904,8 +904,7 @@ nsTableRowFrame::ReflowChildren(nsPresContext*           aPresContext,
         // Reflow the child
         kidReflowState.emplace(aPresContext, aReflowState, kidFrame,
                                kidAvailSize,
-                               // Cast needed for gcc 4.4.
-                               uint32_t(nsHTMLReflowState::CALLER_WILL_INIT));
+                               nsHTMLReflowState::CALLER_WILL_INIT);
         InitChildReflowState(*aPresContext, kidAvailSize, borderCollapse,
                              *kidReflowState);
